+++ libs/core/src/system_state/monitors.rs
@@ -1,6 +1,6 @@
 use crate::{identifier_impl, rect::Rect};
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[cfg_attr(all(feature = "gen-binds", not(feature = "salvo")), derive(ts_rs::TS))]
 #[serde(rename_all = "camelCase")]
 pub struct PhysicalMonitor {